requirements.txt
python-dotenv
google-generativeai
django
vertexai